+++ x/distribution/keeper/migrations.go
@@ -3,7 +3,7 @@ package keeper
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/distribution/exported"
-	v043 "github.com/cosmos/cosmos-sdk/x/distribution/migrations/v043"
+	v2 "github.com/cosmos/cosmos-sdk/x/distribution/migrations/v2"
 	v3 "github.com/cosmos/cosmos-sdk/x/distribution/migrations/v3"
 )
 
@@ -20,7 +20,7 @@ func NewMigrator(keeper Keeper, legacySubspace exported.Subspace) Migrator {
 
 // Migrate1to2 migrates from version 1 to 2.
 func (m Migrator) Migrate1to2(ctx sdk.Context) error {
-	return v043.MigrateStore(ctx, m.keeper.storeKey)
+	return v2.MigrateStore(ctx, m.keeper.storeKey)
 }
 
 // Migrate2to3 migrates the x/distribution module state from the consensus